Check if greater than unsigned long

+++ 0x14-bit_manipulation/2-get_bit.c
@@ -11,7 +11,7 @@ int get_bit(unsigned long int n, unsigned int index)
 {
 	unsigned int count;
 
-	if (n < sizeof(unsigned int))
+	if (n < sizeof(unsigned long int))
 		return (-1);
 
 	for (count = 0; count < index; count++)
